Add files via upload
Adjusted grass tiles with slightly less detail and a toned down green so it doesn't fight with Watt for attention.  I will continue to simplify.  Please drop this in and see how the tiles look.  I'd like a screen cap or to share a screen with you and see what it looks like.